requirements.txt
pandas>=2.0.0
numpy>=1.24.0
ccxt>=4.0.0
colorama>=0.4.6
python-dotenv>=1.0.0

